Don't allow a file to be added to deleting list twice
This could happen if you try and delete a file with a croppa-style suffix.  It would be in the array twice, once as the source, and once for matching the pattern.

Fixes #48

Also, I am explicitly checking that files begin with the passed filename.  And I added a new unit test