Provided a way to bind without Laravel-Flysystem
Hello,

I've simply provided a way to make an IoC binding without the need for the whole Laravel Flysystem package. I wasn't using that package, and I didn't want to install an additional dependency just for binding. 

This way, you won't be needing the whole requirement just for this.

Thanks in advance,