Don’t build Flysystem disks until needed
Instantiating those disks actually created directories in the filesystem.  This caused problems with `php artisan storage:link` where Croppa would create an empty public/storage immediately before the symlink was created.  It was a side effect of the storage class getting booted because it was a dependecy of many classes that get insantiated with artisan.